Modified the interface of "Room"" and keep a ref to a "data source"
This change will simplify the implementation of other back-ends.
The new "src" property will point to the data source, it could be
an in-memory data structure or a database connection.

The caller of the "Room" class owns that the pointer to the data
source, this means that the caller *needs to make sure* that the
data source object lives at least as long as the "Room" object.