the "Guest" class takes a pointer to a "GuestsData" (memory backend)
The concrete approach is that instead of taking a "GuestData" pointer,
it takes a "GuestsData" instead.

The idea behind this change is motivated that currently in the case of
the "Room" implementation, the "Hotel" needs to search for the room and
then pass the "src" for the found room to the constructor; this isn't
good because the "Hotel" _needs_ to know how about the structure inside
the "rooms container", this shouldn't be its responsibility. With this
new approach used in the "Guest" class, we can now apply the change to
the "Rooms" class so it takes the rooms container as source and do the
search by itself; then only "Rooms" will hold the knowledge about the
internals of the rooms container.