the "Reservations" methods expect a pointer to "HotelData" as src
After using this entity it was observed that passing a "HotelData" as
the source, increases the code duplication in  the "Hotel" implementation
in both backends, that the Hotel is now less dependent on the backend
and hence more generic. A further iteration is to remove this redundancy
in the code.