attempt to factor out the logic to formatting the output
It may be the case that the output is intended to a human user,
or maybe to another service so the formatter will take care of
serializing the output, so the application doesn't need to know
about the details of where the output goes to.

This first version of the "Formatter" takes care of the serialization,
but it's very likely that serialization will be made by a third party
library, so it's main task will be to offer an interface between
the application and the different serializers.