the "Reservation" has a new general purpose "notes" field
The idea behind this new field, is that the "receptionist" might want to keep
notes about unpredictable events during the reservation. This space is ideal to
keep track of all the expenses related to the reservation, for example to record
the rates at the time of the reservation, this way it's can keep track of the
original rates even if the room rates changed after the reservation was created.