explicitly set the "checkin_at" and "checkout_at" in the reservation
For now I'm not including it in the default constructor since the default value
of those two properties is `std::nullopt`. I cannot pass it to the constructor
as an optional argument because the last argument of the constructor is the
"src" and this isn't an optional argument. An idea to avoid this problem is to
make the "src" the first argument of the constructor.