make the "in-memory" mockhotel a copy from the unit-test hotel
This change removes duplication of code, and probably in the future
it may be possible to remove the "t_memory.hh" file.

For this change it was necessary to adjust the tests a bit, one example
is that the unit-test hotel had some pre-loaded reservations, whereas
the mock-hotel doesn't have any reservations, so the expected reservations
where created on the fly in the unit test file.