provide a separate "Makefile.inc" for the Python bindings
This file contains the details for building the Python bindings and is meant to
be imported by the main Makefile. It was decided to be done this way because the
targets were triggering warnings about the "pybind11" library not being
installed, and this is _not_ a required dependency; it was quite difficult to go
around this warning without making over-complicated target rules; splitting is
the cleanest solution so far.