update integration test with the expected "Period" serialization
The "Period" stringification from the boost's Gregorian library is misleading
for this purpose; the reason is that a reservation that begins in a given date
for one day, it stringifies as it begins and ends on the same day. This change
is to start implementing a stringifies in a way where a reservation _doesn't_
seem to start and end on the same day.